feat: Added admin feature for uploading CSV and image files, API response

+++ openfield/farm/models.py
@@ -19,8 +19,8 @@ def generate_farm_image_filename(instance, filename):
 class Farm(models.Model):
     farm_id = models.AutoField(primary_key=True)
     farm_owner = models.CharField(max_length=255)
-    latitude = models.IntegerField()
-    longitude = models.IntegerField()
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     farm_name = models.CharField(max_length=255, default='Unknown')  
     farm_size = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
